needed to make sure that the user coords are not null when user uses geolocation to make a post at their current location. I had a check the sent an alert to the user that they
Also needed to fix the IGeoJson model. It has location instead of geometry as a variable and meant it didnt comply with GeoJson object requred in mapbox. SO i have to convert using GeoJson object, which is longer. So updated the mongoDb sschema to have geometry instead of location so everything matches